Write some tests and organize some files
